#435 Revision ValidationRuleController

+++ grails-app/controllers/au/org/ala/volunteer/ValidationRuleController.groovy
@@ -2,16 +2,26 @@ package au.org.ala.volunteer
 
 class ValidationRuleController {
 
+    def userService
+
     def index() {
-        redirect(action:list())
+        redirect(action: list())
     }
 
     def list() {
+        if (!userService.isAdmin()) {
+            redirect(uri: "/")
+            return
+        }
         def validationRules = ValidationRule.list(params)
-        [validationRules: validationRules, totalCount: ValidationRule.count ]
+        [validationRules: validationRules, totalCount: ValidationRule.count]
     }
 
     def delete() {
+        if (!userService.isAdmin()) {
+            redirect(uri: "/")
+            return
+        }
         def rule = ValidationRule.get(params.int("id"));
         if (rule) {
             rule.delete()
@@ -19,20 +29,32 @@ class ValidationRuleController {
         } else {
             flash.message = "No rule with id ${params.id} exists."
         }
-        redirect(action:"list")
+        redirect(action: "list")
     }
 
     def addRule() {
-        def rule = new ValidationRule(name:"<New rule>")
+        if (!userService.isAdmin()) {
+            redirect(uri: "/")
+            return
+        }
+        def rule = new ValidationRule(name: "<New rule>")
         render(view: 'edit', model: [rule: rule])
     }
 
     def edit() {
+        if (!userService.isAdmin()) {
+            redirect(uri: "/")
+            return
+        }
         def rule = ValidationRule.get(params.int("id"))
         [rule: rule]
     }
 
     def update() {
+        if (!userService.isAdmin()) {
+            redirect(uri: "/")
+            return
+        }
         def rule = ValidationRule.get(params.int("id"))
         if (rule) {
             rule.properties = params
@@ -41,11 +63,11 @@ class ValidationRuleController {
         }
 
         if (!rule.save()) {
-            render(view:'edit', model:[rule: rule])
+            render(view: 'edit', model: [rule: rule])
             return
         }
 
-        redirect(action:'list')
+        redirect(action: 'list')
     }
 
 }
